Move string sort tests from TestSorting to NaturalSortTest
